gophercloud: correct handling of LLDP data
The LLDP data in the inventory doesn't appear as a JSON object, but as
a JSON array of [type, value] pairs such as:

[[1, "04112233aabbcc"], [2, "07373334"], ...]

(There can be multiple fields with the same type, which is why they are
not returned as a JSON object.)

There is a place where the raw LLDP data is processed into a JSON object
- under the `lldp_processed` key in the top-level `all_interfaces` list
(i.e. not in the inventory), here represented by BaseInterfaceType.

This patch corrects handling for LLDP data in the inventory (i.e. in
InterfaceType) and adds handling for it in BaseInterfaceType