ensure last updated is set when extracting status from annotation
If the status annotation is a partial version of the status struct it
will not include the LastUpdated value. This results in an infinite
loop, since hostHasStatus() only looks at the LastUpdated value to see
if it is filled in. This change removes the loop condition by ensuring
LastUpdated is set if the value from the annotation is empty.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>
(cherry picked from commit e0ccd87fd0d29f69e0723808a133ff58a60c72f8)