fix saveHostStatus conflict loop issue
In saveHostStatus, we get the BMH right after updating the
annotation. In some cases the cache is not updated yet, causing
an error in the status update. This could lead to a loop in
some cases where the cache was slow to update as the annotations
were always changed since operationHistory was changing every
time (it uses the current time as end time for an operation).
To solve this :

- Discard operationHistory when comparing status annotation to
  current status to prevent unneeded updates of the annotation
  leading to additional conflicts
- Do not throw an error upon conflict in saveHostStatus, but
  requeue after an info message

(cherry picked from commit 871926d407c4758f96a7fc011f254a57b2c0eaa1)