Unbreak linter jobs by pinning versions
Pre-modules versions of "go get" alway get the latest version of a
package, in the completely insane hope that package authors will keep
compatibility for ever. The latest version of gosec is incompatible with
go1.10, so the gate is broken. Check out and build a slightly older
version (v2.3.0) instead. Other dependencies have also broken
compatibility with go1.10, so we have to pin those as well.