Clear ErrorCount independently of ErrorType/Message
When a call to the provisioner succeeds rather than returning an error
message, that's a good sign, and a reason to not have an error message
set in the Host object. But it doesn't guarantee success: if the
previous failure came at the end of some long-running operation (rather
than outright rejection at the beginning), it could yet fail in exactly
the same way.

Clearing the ErrorType as soon as we start an operation allows us to use
that field to determine whether to force the provisioner to retry in the
presence of an error. (It will only be set if the last thing we did was
record an error, therefore if we see it then we are at the beginning of
a new retry.)

One known issue with this is that if a request to change state in Ironic
results in a 409 Conflict response, this just sets the Dirty flag and is
indistinguishable from success when viewed outside the Ironic
provisioner. If such a conflict occurs, we will effectively skip a retry
attempt, increment the ErrorCount again, and sleep for even longer
before the next attempt.

To ensure that we actually do exponential backoff between retries, leave
the ErrorCount unchanged until the whole operation actually succeeds
(Dirty is no longer true). This is now decoupled from the ClearError()
method that clears the ErrorType and ErrorMessage.
As much as possible, do the clearing of ErrorCount in the host state
machine. The exception is the steady states where we only do power
management - the actionResult types currently lack enough detail to
distinguish when count should be cleared when viewed from the state
machine.

In power management states (Ready/Available, Provisioned, Externally
Provisioned), count the number of errors of any type since the power
state was last successfully changed.

Otherwise, the ErrorCount is cleared when an operation succesfully
completes. Successful re-registration or adoption is never sufficient to
clear the error count, except in the registration state.

(cherry picked from commit 562b6ac58e8be1baa5e968beed5d9912a3f8f153)
Signed-off-by: Honza Pokorny <honza@redhat.com>