Return provisioner ID from ValidateManagementAccess()
The provisioner interface currently requires provisioners to write to the
host's Status in order to set the ID. This is undesirable. Return the
provisioner's ID for the host from ValidateManagementAccess() and write it
to the Status in the controller.

(cherry picked from commit 50f97a7558adff7f1c04faf0452bb1c1d1d6fcfd)
Signed-off-by: Honza Pokorny <honza@redhat.com>