Return power status from UpdateHardwareState()
The provisioner interface currently requires provisioners to write to the
host's Status in order to set the power status. This is undesirable. Return
the power status (if known) for the host from UpdateHardwareState(), rather
than a Result structure from which only the Dirty flag was used. Write the
updated power state to the Status in the controller.

(cherry picked from commit cd60f7dfae393e5e6725d736aef7ae25e3d716dc)
Signed-off-by: Honza Pokorny <honza@redhat.com>