Stop treating actionRegistering() like a state action
All of the action<StateName>() methods of the reconciler are the handlers
used by the host state machine to carry out the action in the current
provisioning state, except for actionRegistering(). actionRegistering() is
called in every state (by the state machine's ensureRegistered() method)
prior to calling the action method for the current state.

Rename actionRegistering() to registerHost() to eliminate the implication
that it is an action specific to the Registering state.

Simplify the return handling by returning nil when validating management
access succeeded and there were no changes required to the Status, since
this is what we want to return from ensureRegistered() (indicating no need
to return early before executing the state's action method). This will
occur both in general operation when there is no re-registration required,
and also potentially at the conclusion of a re-registration where there was
no credentials change and errors had already been cleared (by a previous
invocation that returned actionUpdate).

At the conclusion of a registration where changes need to be written to the
Status, e.g. by clearing a previous error state or updating the known good
credentials, actionComplete is returned. This is later translated to
actionUpdate by ensureRegistered().

This enables that scenario to be distinguished from that of a change
needing to be written while registration is still in progress, which
results in actionUpdate.

This means there is no longer a need in ensureRegistered() to guess when
the registerHost() method might have resulted in a change to the Status.

(cherry picked from commit 98d713159488832fcc367e277abf4601bd6ab8f8)
Signed-off-by: Honza Pokorny <honza@redhat.com>