Fail registration when boot MAC conflicts
If we can't find a node by name, and have to resort to using MAC address
lookup, we can't accept a node that has a name.  In that scenario, we
have detected a MAC address collision, and human intervention is
required.  We signal this by returning an `operationFailed` provisioning
result with the offending MAC address in the error message.

(cherry picked from commit 1dc7f4c8b151e650ae9c2478b799b7aa9b9359bd)
Signed-off-by: Honza Pokorny <honza@redhat.com>