Pass BootMode explicitly to provisioner
Don't rely on the BootMode in the BareMetalHost, but instead pass it
explicitly to those Provisioner methods that require it
(ValidateManagementAccess, InspectHardware, and Provision).

This also resolves an inconsistency where we were sometimes looking at
the BootMode in the Spec and sometimes the one in the Status. We now
look only at the one in the Status. To allow this to do something useful
during Registration, also copy the BootMode from the Spec to the Status
as part of entering Registration (previously this did not happen until
entering Inspecting).