Select empty provisioner in controller
Since cadeb28e313602613e8e1944012d153d4b11bafa we have been logging
provisioner information on every reconcile when it was intended to be
logged only once, at startup.

To fix that, move the selection of the empty provisioner into the
controller, rather than main.go. This is the more logical place for it
anyway, since it depends on the state of the Host and not the startup
environment. This also gives more consistent behaviour for the
demo/fixture provisioners, as unmanaged hosts will now always use the
empty provisioner as they would in a real deployment.

Fixes #819