Add the outline of a coreos provisioner
The CoreOS provisioner uses another provisioner (currently ironic) to
boot an ISO image containing an agent (not ironic-python-agent) in
the registration state. From there inspection and provisioning are done
by communicating directly with the agent.

This functionality can be enabled by setting the ASSISTED_INSTALLER_URL
to the URL of an ISO image to boot.

Still to be implemented:
 - Communication with the agent
 - Fetching a host-specific ISO URL from the service at the configured
   URL, rather than using the configured URL directly