ironic: Add code for lazily updating options
Code to generate UpdateOpts to set options values in ironic is
repetitive and verbose, and often lacks niceties such as checking
whether the correct value is already set before initiating an update.

Add a convenience function updateOptionsSection() that produces
UpdateOpts that add or remove the desired values from either the
Properties or InstanceInfo sections, only when needed.