ironic: Add convenience function for top-level options
Take advantage of the infrastructure we have built for lazily setting
the properties and instance_info options to set top-level options. This
simplifies the code and ensures consistent logging.

A new convenience function is added to ensure a simple interface. To
avoid doing reflection on the Node object, we just pass the current
value as a parameter.