DO NOT MERGE Avoid excessive manual cleaning
Previously we always started a manual cleaning in the Preparing state to
run the "delete_configuration" clean step whenever the driver supports
RAID. This can be expensive since, apart from the cost of cleaning
itself, whenever there are clean steps we have to return from the
available state to manageable, and returning to available again requires
automated cleaning.

This change ensures the "delete_configuration" step is added only when
RAID had already previously been configured on the host, or we had a
previous error in preparing.

HOWEVER, we currently rely on resetting the current RAID settings on
deprovisioning to trigger a new Prepare step. If the RAID settings are
cleared at the same time as deprovisioning then this will not happen (a
problem that pre-dates this patch) unless clearing the root device hints
also causes a mismatch, and after this patch it will not happen even if
the root device hints have a mismatch. The RAID settings are not cleared
by an automated clean, and we shouldn't rely on the settings left behind
by the last user of the host.

In addition, when a host is initally created, we don't know the RAID
status so we should probably clear it out even though it may be
expensive (although in this case at least we don't need an extra
automated clean). Previously we were doing this but this patch prevents
it.

These issues could be solved by having a flag in the BareMetalHost
Status to indicate whether the Host has been prepared since the last
time it was provisioned. However, it's not clear that this would leave
us in any different situation than we are today (although it would
remove the ambiguity between the status being cleared because we've just
deprovisioned and the status being cleared because we're about to clear
the RAID settings... though that might already be being accomplished
indirectly by the root device hints).